requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
Django==4.1.5
django-jazzmin==2.6.0
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
